intx: Move to_string() to int128.hpp
