+++ include/intx/intx.hpp
@@ -784,30 +784,6 @@ constexpr uint<N>& operator%=(uint<N>& x, const T& y) noexcept
 }
 
 template <unsigned N>
-inline std::string to_string(uint<N> x, int base = 10)
-{
-    if (base < 2 || base > 36)
-        throw std::invalid_argument{"invalid base: " + std::to_string(base)};
-
-    if (x == 0)
-        return "0";
-
-    auto s = std::string{};
-    while (x != 0)
-    {
-        // TODO: Use constexpr udivrem_1?
-        const auto res = udivrem(x, base);
-        const auto d = int(res.rem);
-        const auto c = d < 10 ? '0' + d : 'a' + d - 10;
-        s.push_back(char(c));
-        x = res.quot;
-    }
-    std::reverse(s.begin(), s.end());
-    return s;
-}
-
-
-template <unsigned N>
 inline uint<N> bswap(const uint<N>& x) noexcept
 {
     return {bswap(x.lo), bswap(x.hi)};
